+++ python/cuml/cuml/metrics/cluster/__init__.py
@@ -1,23 +1,3 @@
-#
-# SPDX-FileCopyrightText: Copyright (c) 2021-2025, NVIDIA CORPORATION.
-# SPDX-License-Identifier: Apache-2.0
-#
+from .adjusted_mutual_info_score import adjusted_mutual_info_score
 
-from cuml.metrics.cluster.adjusted_rand_index import adjusted_rand_score
-from cuml.metrics.cluster.completeness_score import (
-    cython_completeness_score as completeness_score,
-)
-from cuml.metrics.cluster.entropy import cython_entropy as entropy
-from cuml.metrics.cluster.homogeneity_score import (
-    cython_homogeneity_score as homogeneity_score,
-)
-from cuml.metrics.cluster.mutual_info_score import (
-    cython_mutual_info_score as mutual_info_score,
-)
-from cuml.metrics.cluster.silhouette_score import (
-    cython_silhouette_samples as silhouette_samples,
-)
-from cuml.metrics.cluster.silhouette_score import (
-    cython_silhouette_score as silhouette_score,
-)
-from cuml.metrics.cluster.v_measure import cython_v_measure as v_measure_score
+__all__ = ["adjusted_mutual_info_score"]
