--- python/cuml/metrics/cluster/__init__.py
@@ -1,3 +1,0 @@
-from .adjusted_mutual_info_score import adjusted_mutual_info_score
-
-__all__ = ["adjusted_mutual_info_score"]
